Move auto markdown for note links into plugin
