+++ lib/api/upsertNote.ts
@@ -1,8 +1,11 @@
 import { store } from 'lib/store';
 import supabase from 'lib/supabase';
 import { Note } from 'types/supabase';
+import type { PickPartial } from 'types/utils';
 
-export default async function upsertNote(note: Partial<Note>) {
+type UpsertNote = PickPartial<Note, 'id' | 'content'>;
+
+export default async function upsertNote(note: UpsertNote) {
   const { data } = await supabase
     .from<Note>('notes')
     .upsert(note, { onConflict: 'user_id, title' })
