+++ types/utils.d.ts
@@ -1,0 +1,2 @@
+export type PickPartial<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>> &
+  Partial<Pick<T, K>>;
